feat(EventManager): add toggle item event listener
This commit introduces a new event listener for the toggle item in the popup menu. It also updates the 'createPopupMenuElement' method in the 'DOMManager' class to set the text content of the toggle item based on the collapse status of the message. The 'addMessage' method in the 'MessageManager' class is also updated to create the popup menu with the correct collapse status when the page loads.